Fix AttachCapabilitiesEvent applying to non-worlds
